requirements.txt
fastapi
uvicorn[standard]
openai-whisper
python-multipart
jinja2
deep-translator
